requirements.txt
appdirs==1.4.3
astroid==2.0.4
attrs==18.2.0
black==18.9b0
Cerberus==1.2
Click==7.0
Eve==0.8.1
Events==0.3
Flask==1.0.2
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
logzero==1.5.0
MarkupSafe==1.1.0
mccabe==0.6.1
numpy==1.15.3
paho-mqtt==1.4.0
pandas==0.23.4
pluginbase==0.7
pylint==2.1.1
pymongo==3.7.2
python-dateutil==2.7.5
pytz==2018.7
simplejson==3.16.0
six==1.11.0
toml==0.10.0
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
